CATALOG-3148: Move table_function trait from catalogs to table_functions crate
