+++ query/src/table_functions/mod.rs
@@ -21,9 +21,11 @@ mod numbers_table_test;
 mod memory_block_part;
 mod numbers_stream;
 mod numbers_table;
+mod table_function;
 mod table_function_factory;
 
 pub use memory_block_part::generate_block_parts;
 pub use numbers_table::NumbersTable;
+pub use table_function::TableFunction;
 pub use table_function_factory::TableArgs;
 pub use table_function_factory::TableFunctionFactory;
